Ещё пакет правок

See merge request gashiykh/laboratory_work_64_team_3!32

+++ instagram/webapp/views/like.py
@@ -1,5 +1,6 @@
 from django.http import JsonResponse
 from django.shortcuts import get_object_or_404
+from django.urls import reverse
 from django.views import View
 
 from webapp.models import Post
@@ -9,22 +10,31 @@ from webapp.models import Like
 class LikeView(View):
     def get(self, request, post_id):
         user = request.user
-        post = get_object_or_404(Post, id=post_id)
-        liked = Like.objects.filter(author=user, post=post).exists()
+        if user.is_authenticated:
+            post = get_object_or_404(Post, id=post_id)
+            liked = Like.objects.filter(author=user, post=post).exists()
 
-        if liked:
-            Like.objects.filter(author=user, post=post).delete()
-            post.like_count -= 1
-            liked = False
+            if liked:
+                Like.objects.filter(author=user, post=post).delete()
+                post.like_count -= 1
+                liked = False
+            else:
+                Like.objects.create(author=user, post=post)
+                post.like_count += 1
+                liked = True
+
+            post.save()
+
+            response = {
+                'success': True,
+                'liked': liked,
+                'like_count': post.like_count
+            }
         else:
-            Like.objects.create(author=user, post=post)
-            post.like_count += 1
-            liked = True
+            response = {
+                'success': False,
+                'error': 'login_required',
+                'login_url': reverse('login')
+            }
 
-        post.save()
-
-        return JsonResponse({
-            'success': True,
-            'liked': liked,
-            'like_count': post.like_count
-        })
+        return JsonResponse(response)
